+++ src/main/java/subway/application/AddSectionService.java
@@ -3,8 +3,8 @@ package subway.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.application.dto.AddSectionDto;
-import subway.domain.Direction;
-import subway.domain.Distance;
+import subway.domain.section.Direction;
+import subway.domain.section.Distance;
 import subway.domain.line.Line;
 import subway.domain.station.Station;
 import subway.persistence.repository.LineRepository;
