+++ src/main/java/subway/domain/section/Sections.java
@@ -9,8 +9,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Queue;
 import java.util.Set;
-import subway.domain.Direction;
-import subway.domain.Distance;
 import subway.domain.station.Station;
 
 public class Sections {
