+++ src/main/java/subway/persistence/repository/SectionRepository.java
@@ -9,8 +9,8 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
-import subway.domain.Direction;
-import subway.domain.Distance;
+import subway.domain.section.Direction;
+import subway.domain.section.Distance;
 import subway.domain.line.Line;
 import subway.domain.section.Section;
 import subway.domain.section.Sections;
