+++ src/main/java/subway/ui/dto/request/CreationSectionRequest.java
@@ -2,7 +2,7 @@ package subway.ui.dto.request;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
-import subway.domain.Direction;
+import subway.domain.section.Direction;
 
 public class CreationSectionRequest {
